+++ source/basic_locomotion_isaaclab/basic_locomotion_isaaclab/tasks/locomotion/agents/morphosymm_cfg.py
@@ -1,0 +1,79 @@
+from isaaclab.utils import configclass
+
+from pathlib import Path
+from dataclasses import MISSING
+
+@configclass
+class MorphologycalSymmetriesCfg:
+    """Configuration for using morphosymm-rl."""
+
+    class_name: str = "MorphologycalSymmetries"
+    """The class name."""
+
+    obs_space_names_actor =  None
+    """The observation space names for the actor network."""
+
+    obs_space_names_critic = None
+    """The observation space names for the critic network."""
+
+    action_space_names = None
+    """The action space names."""
+
+    joints_order = None
+    """The order of the joints in the robot."""
+
+    robot_name = None
+    """The name of the robot to use inside Morphosymm."""
+
+
+# Actor OBS
+history_length = 5
+obs_space_names_actor = [
+        "base_lin_vel:base",
+        "base_ang_vel:base",
+        "gravity:base",
+        "ctrl_commands",
+        "default_qpos_js_error",
+        "qvel_js",
+        "actions",
+        "clock_data",
+    ]*int(history_length)
+
+
+# Critic OBS
+obs_space_names_critic = [
+        "base_lin_vel:base",
+        "base_ang_vel:base",
+        "gravity:base",
+        "ctrl_commands",
+        "default_qpos_js_error",
+        "qvel_js",
+        "actions",
+        "clock_data",
+    ]*int(history_length)
+obs_space_names_critic += ["base_pos_z", "base_pos_z", "clock_data", "default_qpos_js_error", "default_qpos_js_error"]
+
+
+# Action Space
+action_space_names = ["actions"]
+
+
+# Joints Order
+joints_order = [
+    "FL_hip_joint", "FR_hip_joint", "RL_hip_joint", "RR_hip_joint", 
+    "FL_thigh_joint", "FR_thigh_joint", "RL_thigh_joint", "RR_thigh_joint",
+    "FL_calf_joint", "FR_calf_joint", "RL_calf_joint", "RR_calf_joint"
+]
+
+
+# Robot Name
+robot_name = "a1"
+
+
+morphologycal_symmetries_cfg = MorphologycalSymmetriesCfg(
+        obs_space_names_actor = obs_space_names_actor,
+        obs_space_names_critic = obs_space_names_critic,
+        action_space_names = action_space_names,
+        joints_order = joints_order,
+        robot_name = robot_name,
+    )
