+++ source/basic_locomotion_isaaclab/basic_locomotion_isaaclab/tasks/locomotion/agents/rsl_rl_ppo_cfg.py
@@ -7,50 +7,9 @@
 from isaaclab.utils import configclass
 from isaaclab_rl.rsl_rl import RslRlOnPolicyRunnerCfg, RslRlPpoActorCriticCfg, RslRlPpoAlgorithmCfg
 
-from pathlib import Path
-from dataclasses import MISSING
-
-@configclass
-class DiscriminatorCfg:
-    """Configuration for the discriminator network."""
-
-    class_name: str = "Discriminator"
-    """The discriminator class name. Default is Discriminator."""
-
-    hidden_dims: list[int] = MISSING
-    """The hidden dimensions of the discriminator network."""
-
-    reward_scale: float = MISSING
-    """The reward coefficient."""
-
-    loss_type: str = "BCEWithLogits"
-    """The type of loss to use for training the discriminator. Default is BCEWithLogits."""
-    
-    empirical_normalization: bool = False
-    """Whether to use empirical normalization for the discriminator inputs. Default is False."""
-
-
-@configclass
-class MorphologycalSymmetriesCfg:
-    """Configuration for using morphosymm-rl."""
-
-    class_name: str = "MorphologycalSymmetries"
-    """The class name."""
-
-    obs_space_names_actor =  None
-    """The observation space names for the actor network."""
-
-    obs_space_names_critic = None
-    """The observation space names for the critic network."""
-
-    action_space_names = None
-    """The action space names."""
-
-    joints_order = None
-    """The order of the joints in the robot."""
-
-    robot_name = None
-    """The name of the robot to use inside Morphosymm."""
+from copy import deepcopy
+from . import amp_cfg
+from . import morphosymm_cfg
 
 
 @configclass
@@ -83,51 +42,15 @@ class FlatPPORunnerCfg(RslRlOnPolicyRunnerCfg):
         max_grad_norm=1.0,
     )
 
-    #AMP Related Stuff
-    amp_data_path = "./../../../../../../scripts/amp_rl/amp_dataset/"
-    dataset_names = ["flat"]
-    dataset_weights = [1.0, 1.0]
-    slow_down_factor = 1.0
-    discriminator = DiscriminatorCfg(
-        hidden_dims=[128, 128],
-        reward_scale=0.1,
-        loss_type="BCEWithLogits",
-        empirical_normalization= False
-    )
+    # AMP Related Stuff
+    amp_data_path = amp_cfg.amp_data_path
+    dataset_names = list(amp_cfg.dataset_names)
+    dataset_weights = list(amp_cfg.dataset_weights)
+    slow_down_factor = amp_cfg.slow_down_factor
+    discriminator = deepcopy(amp_cfg.discriminator)
 
-    # Symmetry Related Stuff - Actor Critic
-    history_length = 5
-    obs_space_names_actor = [
-            "base_lin_vel:base",
-            "base_ang_vel:base",
-            "gravity:base",
-            "ctrl_commands",
-            "default_qpos_js_error",
-            "qvel_js",
-            "actions",
-            "clock_data",
-        ]*int(history_length)
-    obs_space_names_critic = obs_space_names_actor
-    
-    # Symmetry Related Stuff -  Asymmetric Critic
-    """obs_space_names_critic += ["position_gains", 
-            "velocity_gains",
-            "friction_static",
-            "friction_dynamic",
-            "armature"
-        ]"""
-
-    morphologycal_symmetries_cfg = MorphologycalSymmetriesCfg(
-        obs_space_names_actor = obs_space_names_actor,
-        obs_space_names_critic = obs_space_names_critic,
-        action_space_names = ["actions"],
-        joints_order = [
-            "FL_hip_joint", "FR_hip_joint", "RL_hip_joint", "RR_hip_joint", 
-            "FL_thigh_joint", "FR_thigh_joint", "RL_thigh_joint", "RR_thigh_joint",
-            "FL_calf_joint", "FR_calf_joint", "RL_calf_joint", "RR_calf_joint"
-        ],
-        robot_name = "a1",
-    )
+    # Morphosymm-rl Related Stuff
+    morphologycal_symmetries_cfg = morphosymm_cfg.morphologycal_symmetries_cfg
 
 
 @configclass
@@ -140,8 +63,6 @@ class RoughPPORunnerCfg(RslRlOnPolicyRunnerCfg):
     policy = RslRlPpoActorCriticCfg(
         class_name="ActorCritic", #ActorCritic, ActorCriticRecurrent, ActorCriticSymm, ActorCriticMoE
         init_noise_std=1.0,
-        #actor_hidden_dims=[512, 256, 128],
-        #critic_hidden_dims=[512, 256, 128],
         actor_hidden_dims=[128, 128, 128],
         critic_hidden_dims=[128, 128, 128],
         activation="elu",
@@ -162,57 +83,12 @@ class RoughPPORunnerCfg(RslRlOnPolicyRunnerCfg):
         max_grad_norm=1.0,
     )
 
-    #AMP Related Stuff
-    amp_data_path = "./../../../../../../scripts/amp_rl/amp_dataset/"
-    dataset_names = ["flat", "boxes", "stairs"]
-    dataset_weights = [1.0, 1.0, 1.0, 1.0]
-    slow_down_factor = 1.0
-    discriminator = DiscriminatorCfg(
-        hidden_dims=[1024, 512],
-        reward_scale=1.0,
-        loss_type="BCEWithLogits",
-        empirical_normalization= False
-    )
+     # AMP Related Stuff
+    amp_data_path = amp_cfg.amp_data_path
+    dataset_names = list(amp_cfg.dataset_names)
+    dataset_weights = list(amp_cfg.dataset_weights)
+    slow_down_factor = amp_cfg.slow_down_factor
+    discriminator = deepcopy(amp_cfg.discriminator)
 
-    # Symmetry Related Stuff - Actor Critic
-    history_length = 5
-    obs_space_names_actor = [
-            "base_lin_vel:base",
-            "base_ang_vel:base",
-            "gravity:base",
-            "ctrl_commands",
-            "default_qpos_js_error",
-            "qvel_js",
-            "actions",
-            "clock_data",
-        ]*int(history_length)
-
-    #obs_space_names_actor += ["heightmap:rows4xcols4"]
-    #obs_space_names_actor += ["position_gains"]
-    #obs_space_names_actor += ["velocity_gains"]
-
-
-    # Symmetry Related Stuff -  Asymmetric Critic
-    obs_space_names_critic = [
-            "base_lin_vel:base",
-            "base_ang_vel:base",
-            "gravity:base",
-            "ctrl_commands",
-            "default_qpos_js_error",
-            "qvel_js",
-            "actions",
-            "clock_data",
-        ]*int(history_length)
-    obs_space_names_critic += ["base_pos_z", "base_pos_z"]
-
-    morphologycal_symmetries_cfg = MorphologycalSymmetriesCfg(
-        obs_space_names_actor = obs_space_names_actor,
-        obs_space_names_critic = obs_space_names_critic,
-        action_space_names = ["actions"],
-        joints_order = [
-            "FL_hip_joint", "FR_hip_joint", "RL_hip_joint", "RR_hip_joint", 
-            "FL_thigh_joint", "FR_thigh_joint", "RL_thigh_joint", "RR_thigh_joint",
-            "FL_calf_joint", "FR_calf_joint", "RL_calf_joint", "RR_calf_joint"
-        ],
-        robot_name = "a1",
-    )
+    # Morphosymm-rl Related Stuff
+    morphologycal_symmetries_cfg = morphosymm_cfg.morphologycal_symmetries_cfg
